requirements.txt
streamlit==1.49.1
tensorflow>=2.20.0
keras==3.10.0
Pillow==11.3.0
matplotlib==3.10.0
